websocket API
